feat(shop-the-look): begin refactor by isolating state and data logic
This commit marks the completion of Phase 1 of the "Shop the Look" refactoring plan.

The primary changes include:
- Creating a dedicated `PageState` class in `state/shop_the_look_state.py` to decouple state management from the UI.
- Creating a new `models/shop_the_look_workflow.py` module to house all data access and storage logic.
- Moving all Firestore-related functions (e.g., `load_model_data`, `store_article_data`) from `pages/shop_the_look.py` to the new workflow module.
- Updating the `pages/shop_the_look.py` file to import the new state and call the data functions from their new location.

This foundational step separates concerns and prepares the feature for further componentization, as outlined in the refactoring plan.
