+++ experiments/veo-app/pages/shop_the_look.py
@@ -26,6 +26,7 @@ import mesop as me
 import requests
 from google.cloud import firestore
 
+import models.shop_the_look_workflow as shop_the_look_workflow
 from common.metadata import MediaItem, add_media_item_to_firestore
 from common.storage import (
     download_from_gcs,
@@ -41,7 +42,6 @@ from components.page_scaffold import (
 )
 from components.tab_nav import Tab, tab_group
 from config.default import Default
-from config.firebase_config import FirebaseClient
 from models.gemini import (
     describe_images_and_look,
     final_image_critic,
@@ -57,82 +57,17 @@ from models.shop_the_look_models import (
 from models.veo import image_to_video
 from models.vto import call_virtual_try_on
 from pages.styles import _BOX_STYLE_CENTER_DISTRIBUTED
+from state.shop_the_look_state import PageState
 from state.state import AppState
 
 GROUP_ORDER = ["workflow-retail-modes"]
 config = Default()
 
 
-@me.stateclass
-class PageState:
-    """Mesop Page State"""
-
-    # TAB NAV
-    selected_tab_index: int = 0
-    disabled: bool = False
-    disabled_tab_indexes: set[int] = field(default_factory=lambda: {-1})
-    mode = "/workflows-retail/lool"
-
-    aspect_ratio: str = "9:16"
-    video_length: int = 8  # 5-8
-
-    # I2V reference Image
-    reference_image_file_clothing: me.UploadedFile = None
-    reference_image_file_key_clothing: int = 0
-    reference_image_gcs_clothing: list[str] = field(default_factory=list)
-    reference_image_uri_clothing: list[str] = field(default_factory=list)
-
-    reference_image_file_model: me.UploadedFile = None
-    reference_image_file_key_model: int = 0
-    reference_image_gcs_model: str
-    reference_image_uri_model: str
-
-    is_loading: bool = False
-    show_error_dialog: bool = False
-    error_message: str = ""
-    result_image: str
-    timing: str
-    look: int = 0
-    catalog: list[CatalogRecord] = field(default_factory=list)
-    models: list[ModelRecord] = field(default_factory=list)
-    before_image_uri: str
-    normal_accordion: dict[str, bool] = field(
-        default_factory=lambda: {
-            "retry_progression": True,
-            "progression": True,
-            "alternate": True,
-        }
-    )
-    progression_images: list[ProgressionImages] = field(default_factory=list)
-    retry_progression_images: list[ProgressionImages] = field(default_factory=list)
-    alternate_progression_images: list[str] = field(default_factory=list)
-    alternate_images: list[str] = field(default_factory=list)
-    veo_prompt_input: str = (
-        "Wide angle shot from a high-angle ceiling perspective captures a sleek model confidently striding down a brightly lit runway. Her full figure is elegantly presented, with every detail of her avant-garde ensemble visible, but the primary focus is drawn to her meticulously designed footwear. The shoes, perhaps gleaming architectural platforms or intricately embellished heels, are highlighted by the stark, dramatic spotlights illuminating the pristine runway below. The long, clean lines of the catwalk stretch into the distance, with the blurred, indistinct forms of the audience fading into the background, ensuring all attention remains on the model's powerful stride and the striking statement of her shoes. The elevated viewpoint offers a unique, almost abstract, composition that emphasizes the geometry of the runway and the singular importance of the footwear. The camera shot should be from 20 feet away."
-    )
-    result_video: str
-    generate_alternate_views: bool
-    selected_model: ModelRecord
-    generate_video: bool = True
-    result_images: list[str] = field(default_factory=list)
-    veo_model = "2.0"
-    current_status: str = ""
-    vto_sample_count: str = "4"
-    veo_sample_count: str = "2"
-    look_description: str = ""
-    final_accuracy: bool
-    final_critic: GeneratedImageAccuracyWrapper = None
-    tryon_started: bool = False
-    articles: list[CatalogRecord] = field(default_factory=list)
-    retry_counter: int = 0
-    max_retry: str = "3"
-    upload_everyone: bool = False
-
-
 @me.component
 def tab_config():
     state = me.state(PageState)
-    state.models = load_model_data()
+    state.models = shop_the_look_workflow.load_model_data()
     vto_enterprise_config()
 
 
@@ -140,10 +75,10 @@ def tab_config():
 def tab_stl():
     state = me.state(PageState)
     if not state.catalog:
-        state.catalog = load_look_data()
+        state.catalog = shop_the_look_workflow.load_look_data()
     if not state.models:
-        state.models = load_model_data()
-        load_article_data()
+        state.models = shop_the_look_workflow.load_model_data()
+        shop_the_look_workflow.load_article_data()
     if state.reference_image_gcs_model is None or not state.reference_image_gcs_model:
         stl_model_select()
     elif not state.look or state.look == 0:
@@ -177,7 +112,6 @@ def on_click_veo(e: me.ClickEvent):  # pylint: disable=unused-argument
     current_error_message = ""
 
     try:
-
         veo_prompt = state.veo_prompt_input
 
         # Attach products descriptions to Veo prompt
@@ -387,41 +321,47 @@ def stl_model_select():
                     #         background="#FFFFFF",
                     #     ),
                     # )
-                    me.icon(
-                        "add",
-                        style=me.Style(
-                            color="black",
-                            position="absolute",
-                            top="35px",
-                            left="35px",
-                            width="35px",
-                            height="35px",
-                            font_size="35px",
-                        ),  # Consider using max_width or responsive units
-                    ),
-                    me.icon(
-                        "person_outline",
-                        style=me.Style(
-                            color="black",
-                            position="absolute",
-                            top="50px",
-                            left="55px",
-                            width="100px",
-                            height="100px",
-                            font_size="100px",
-                        ),  # Consider using max_width or responsive units
-                    ),
-                    me.text(
-                        text="Add Model",
-                        style=me.Style(
-                            position="absolute",
-                            # width="100%",
-                            text_align="center",
-                            bottom="45px",
-                            left="55px",
-                            font_size="20px",
+                    (
+                        me.icon(
+                            "add",
+                            style=me.Style(
+                                color="black",
+                                position="absolute",
+                                top="35px",
+                                left="35px",
+                                width="35px",
+                                height="35px",
+                                font_size="35px",
+                            ),  # Consider using max_width or responsive units
                         ),
-                    ),
+                    )
+                    (
+                        me.icon(
+                            "person_outline",
+                            style=me.Style(
+                                color="black",
+                                position="absolute",
+                                top="50px",
+                                left="55px",
+                                width="100px",
+                                height="100px",
+                                font_size="100px",
+                            ),  # Consider using max_width or responsive units
+                        ),
+                    )
+                    (
+                        me.text(
+                            text="Add Model",
+                            style=me.Style(
+                                position="absolute",
+                                # width="100%",
+                                text_align="center",
+                                bottom="45px",
+                                left="55px",
+                                font_size="20px",
+                            ),
+                        ),
+                    )
 
             for model in state.models:
                 with me.box(
@@ -607,7 +547,6 @@ def stl_result():
                 flex_grow=1,
             )
         ):
-
             me.text(
                 text="Your Look",
                 type="headline-4",
@@ -704,7 +643,6 @@ def stl_result():
                         align_items="center",
                     )
                 ):
-
                     if state.final_critic:
                         with me.box(
                             style=me.Style(
@@ -816,7 +754,6 @@ def stl_result():
                                     align_items="right",
                                 )
                             ):
-
                                 for img in state.alternate_images:
                                     image_url = img.replace(
                                         "gs://",
@@ -1162,41 +1099,47 @@ def stl_look_select():
                         key="top",
                         multiple=True,
                     )
-                    me.icon(
-                        "add",
-                        style=me.Style(
-                            color="black",
-                            position="absolute",
-                            top="20px",
-                            left="30px",
-                            width="30px",
-                            height="30px",
-                            font_size="30px",
-                        ),  # Consider using max_width or responsive units
-                    ),
-                    me.icon(
-                        "apparel",
-                        style=me.Style(
-                            color="black",
-                            position="absolute",
-                            top="45px",
-                            left="55px",
-                            width="50px",
-                            height="50px",
-                            font_size="50px",
-                        ),  # Consider using max_width or responsive units
-                    ),
-                    me.text(
-                        text="Add Top",
-                        style=me.Style(
-                            position="absolute",
-                            # width="100%",
-                            text_align="center",
-                            bottom="35px",
-                            left="45px",
-                            font_size="20px",
+                    (
+                        me.icon(
+                            "add",
+                            style=me.Style(
+                                color="black",
+                                position="absolute",
+                                top="20px",
+                                left="30px",
+                                width="30px",
+                                height="30px",
+                                font_size="30px",
+                            ),  # Consider using max_width or responsive units
                         ),
-                    ),
+                    )
+                    (
+                        me.icon(
+                            "apparel",
+                            style=me.Style(
+                                color="black",
+                                position="absolute",
+                                top="45px",
+                                left="55px",
+                                width="50px",
+                                height="50px",
+                                font_size="50px",
+                            ),  # Consider using max_width or responsive units
+                        ),
+                    )
+                    (
+                        me.text(
+                            text="Add Top",
+                            style=me.Style(
+                                position="absolute",
+                                # width="100%",
+                                text_align="center",
+                                bottom="35px",
+                                left="45px",
+                                font_size="20px",
+                            ),
+                        ),
+                    )
             with me.box(
                 style=me.Style(
                     position="relative",
@@ -1226,41 +1169,47 @@ def stl_look_select():
                     key="bottom",
                     multiple=True,
                 )
-                me.icon(
-                    "add",
-                    style=me.Style(
-                        color="black",
-                        position="absolute",
-                        top="20px",
-                        left="30px",
-                        width="30px",
-                        height="30px",
-                        font_size="30px",
-                    ),  # Consider using max_width or responsive units
-                ),
-                me.icon(
-                    "styler",
-                    style=me.Style(
-                        color="black",
-                        position="absolute",
-                        top="45px",
-                        left="55px",
-                        width="50px",
-                        height="50px",
-                        font_size="50px",
-                    ),  # Consider using max_width or responsive units
-                ),
-                me.text(
-                    text="Add Bottom",
-                    style=me.Style(
-                        position="absolute",
-                        # width="100%",
-                        text_align="center",
-                        bottom="35px",
-                        left="30px",
-                        font_size="20px",
+                (
+                    me.icon(
+                        "add",
+                        style=me.Style(
+                            color="black",
+                            position="absolute",
+                            top="20px",
+                            left="30px",
+                            width="30px",
+                            height="30px",
+                            font_size="30px",
+                        ),  # Consider using max_width or responsive units
                     ),
-                ),
+                )
+                (
+                    me.icon(
+                        "styler",
+                        style=me.Style(
+                            color="black",
+                            position="absolute",
+                            top="45px",
+                            left="55px",
+                            width="50px",
+                            height="50px",
+                            font_size="50px",
+                        ),  # Consider using max_width or responsive units
+                    ),
+                )
+                (
+                    me.text(
+                        text="Add Bottom",
+                        style=me.Style(
+                            position="absolute",
+                            # width="100%",
+                            text_align="center",
+                            bottom="35px",
+                            left="30px",
+                            font_size="20px",
+                        ),
+                    ),
+                )
 
             with me.box(
                 style=me.Style(
@@ -1291,41 +1240,47 @@ def stl_look_select():
                     key="dress",
                     multiple=True,
                 )
-                me.icon(
-                    "add",
-                    style=me.Style(
-                        color="black",
-                        position="absolute",
-                        top="20px",
-                        left="30px",
-                        width="30px",
-                        height="30px",
-                        font_size="30px",
-                    ),  # Consider using max_width or responsive units
-                ),
-                me.icon(
-                    "girl",
-                    style=me.Style(
-                        color="black",
-                        position="absolute",
-                        top="35px",
-                        left="45px",
-                        width="70px",
-                        height="70px",
-                        font_size="70px",
-                    ),  # Consider using max_width or responsive units
-                ),
-                me.text(
-                    text="Add Dress",
-                    style=me.Style(
-                        position="absolute",
-                        # width="100%",
-                        text_align="center",
-                        bottom="35px",
-                        left="35px",
-                        font_size="20px",
+                (
+                    me.icon(
+                        "add",
+                        style=me.Style(
+                            color="black",
+                            position="absolute",
+                            top="20px",
+                            left="30px",
+                            width="30px",
+                            height="30px",
+                            font_size="30px",
+                        ),  # Consider using max_width or responsive units
                     ),
-                ),
+                )
+                (
+                    me.icon(
+                        "girl",
+                        style=me.Style(
+                            color="black",
+                            position="absolute",
+                            top="35px",
+                            left="45px",
+                            width="70px",
+                            height="70px",
+                            font_size="70px",
+                        ),  # Consider using max_width or responsive units
+                    ),
+                )
+                (
+                    me.text(
+                        text="Add Dress",
+                        style=me.Style(
+                            position="absolute",
+                            # width="100%",
+                            text_align="center",
+                            bottom="35px",
+                            left="35px",
+                            font_size="20px",
+                        ),
+                    ),
+                )
 
             with me.box(
                 style=me.Style(
@@ -1356,41 +1311,47 @@ def stl_look_select():
                     key="shoe",
                     multiple=True,
                 )
-                me.icon(
-                    "add",
-                    style=me.Style(
-                        color="black",
-                        position="absolute",
-                        top="20px",
-                        left="30px",
-                        width="30px",
-                        height="30px",
-                        font_size="30px",
-                    ),  # Consider using max_width or responsive units
-                ),
-                me.icon(
-                    "steps",
-                    style=me.Style(
-                        color="black",
-                        position="absolute",
-                        top="45px",
-                        left="55px",
-                        width="50px",
-                        height="50px",
-                        font_size="50px",
-                    ),  # Consider using max_width or responsive units
-                ),
-                me.text(
-                    text="Add Shoe",
-                    style=me.Style(
-                        position="absolute",
-                        # width="100%",
-                        text_align="center",
-                        bottom="35px",
-                        left="35px",
-                        font_size="20px",
+                (
+                    me.icon(
+                        "add",
+                        style=me.Style(
+                            color="black",
+                            position="absolute",
+                            top="20px",
+                            left="30px",
+                            width="30px",
+                            height="30px",
+                            font_size="30px",
+                        ),  # Consider using max_width or responsive units
                     ),
-                ),
+                )
+                (
+                    me.icon(
+                        "steps",
+                        style=me.Style(
+                            color="black",
+                            position="absolute",
+                            top="45px",
+                            left="55px",
+                            width="50px",
+                            height="50px",
+                            font_size="50px",
+                        ),  # Consider using max_width or responsive units
+                    ),
+                )
+                (
+                    me.text(
+                        text="Add Shoe",
+                        style=me.Style(
+                            position="absolute",
+                            # width="100%",
+                            text_align="center",
+                            bottom="35px",
+                            left="35px",
+                            font_size="20px",
+                        ),
+                    ),
+                )
 
             for item in state.articles:
                 with me.box(
@@ -1526,69 +1487,12 @@ def on_blur_veo_prompt(e: me.InputBlurEvent):
 
 
 # End Config Events
-def article_on_delete(e: me.ClickEvent):
-    state = me.state(PageState)
-    # selected_type = e.key.split("-")[-1]
-    # selected_id = e.key.split("-")[-2]
-    file_to_delete = e.key.split("/")[
-        -1
-    ]  # e.key.replace("https://storage.mtls.cloud.google.com/", "gs://")
-    print(f"deleting {file_to_delete}")
-    state.current_status = f"Deleting article {file_to_delete}"
-    try:
-        app_state = me.state(AppState)
-        current_datetime = datetime.datetime.now()
-        doc_ref = db.collection(config.GENMEDIA_VTO_CATALOG_COLLECTION_NAME).document(
-            file_to_delete
-        )
-
-        doc_ref.delete()
-        load_article_data()
-        state.current_status = ""
-        yield
-    except:
-        print(f"Model data  delete failure: {file_to_delete} cannot be stored")
-
-
-def model_on_delete(e: me.ClickEvent):
-    state = me.state(PageState)
-    # selected_type = e.key.split("-")[-1]
-    # selected_id = e.key.split("-")[-2]
-    file_to_delete = e.key.split("/")[
-        -1
-    ]  # e.key.replace("https://storage.mtls.cloud.google.com/", "gs://")
-    print(f"deleting {file_to_delete}")
-    state.current_status = f"Deleting model {file_to_delete}"
-    try:
-        app_state = me.state(AppState)
-        current_datetime = datetime.datetime.now()
-        doc_ref = db.collection(config.GENMEDIA_VTO_MODEL_COLLECTION_NAME).document(
-            file_to_delete
-        )
-
-        doc_ref.delete()
-        state.models = load_model_data()
-        state.current_status = ""
-        yield
-    except:
-        print(f"Model data  delete failure: {file_to_delete} cannot be stored")
 
 
 def vto_enterprise_config():
     state = me.state(PageState)
     app_state = me.state(AppState)
-    # with me.box(
-    #     style=me.Style(
-    #         # flex_basis="450px",
-    #         flex_basis="max(480px, calc(40% - 48px))",
-    #         display="flex",
-    #         flex_direction="column",
-    #         align_items="stretch",
-    #         justify_content="space-between",
-    #         gap=10,
-    #         margin=me.Margin(top=10),
-    #     )
-    # ):
+
     with me.box():
         with me.box(
             style=me.Style(
@@ -1599,7 +1503,6 @@ def vto_enterprise_config():
                 flex_wrap="wrap",
             )
         ):
-
             with me.box(
                 style=me.Style(
                     display="flex",
@@ -1779,7 +1682,7 @@ def vto_enterprise_config():
                         ):
                             with me.box(
                                 key=f"{img}",
-                                on_click=article_on_delete,
+                                on_click=shop_the_look_workflow.article_on_delete,
                                 style=me.Style(
                                     cursor="pointer",
                                 ),
@@ -1861,7 +1764,6 @@ def vto_enterprise_config():
             print(state.models)
             if state.models:
                 for model in state.models:
-
                     with me.box(
                         style=me.Style(
                             display="flex",
@@ -1884,7 +1786,7 @@ def vto_enterprise_config():
                         ):
                             with me.box(
                                 key=f"{img}",
-                                on_click=model_on_delete,
+                                on_click=shop_the_look_workflow.model_on_delete,
                                 style=me.Style(
                                     cursor="pointer",
                                 ),
@@ -1957,58 +1859,6 @@ def on_tab_click(e: me.ClickEvent):
     state.selected_tab_index = int(tab_index)
 
 
-def store_model_data(file_path):
-    # try:
-    state = me.state(PageState)
-    app_state = me.state(AppState)
-    current_datetime = datetime.datetime.now()
-    file_name_only = file_path.split("/")[-1]
-    doc_ref = db.collection(config.GENMEDIA_VTO_MODEL_COLLECTION_NAME)
-
-    upload_user = "everyone" if state.upload_everyone else app_state.user_email
-
-    new_doc_data = {
-        "model_group": "0",
-        "model_id": file_name_only,
-        "model_image": file_path,
-        "timestamp": current_datetime,  # alt: firestore.SERVER_TIMESTAMP
-        "upload_user": upload_user,
-    }
-
-    update_time, doc_ref = doc_ref.add(new_doc_data)
-
-    # except:
-    #     print(f"Model data failure: {file_path} cannot be stored")
-
-
-def store_article_data(file_path, article_category):
-    # try:
-    state = me.state(PageState)
-    app_state = me.state(AppState)
-    current_datetime = datetime.datetime.now()
-    file_name_only = file_path.split("/")[-1]
-    doc_ref = db.collection(config.GENMEDIA_VTO_CATALOG_COLLECTION_NAME)
-
-    upload_user = "everyone" if state.upload_everyone else app_state.user_email
-
-    new_doc_data = {
-        "item_id": file_name_only,
-        "article_type": article_category,
-        "model_group": "0",
-        "timestamp": current_datetime,  # alt: firestore.SERVER_TIMESTAMP
-        "ai_description": None,
-        "selected": False,
-        "available_to_select": True,
-        "clothing_image": file_path,
-        "upload_user": upload_user,
-    }
-
-    update_time, doc_ref = doc_ref.add(new_doc_data)
-
-    # except:
-    #     print(f"Article data failure: {file_path} cannot be stored")
-
-
 def on_click_upload_image(e: me.UploadEvent):
     """Upload image to GCS"""
     state = me.state(PageState)
@@ -2029,7 +1879,7 @@ def on_click_upload_image(e: me.UploadEvent):
                 file.mime_type,
                 file.getvalue(),
             )
-            store_model_data(file_path)
+            shop_the_look_workflow.store_model_data(file_path)
         else:
             file_path = (
                 f"gs://{config.GENMEDIA_BUCKET}/uploads/apparel/{e.key}/{filename}"
@@ -2048,7 +1898,7 @@ def on_click_upload_image(e: me.UploadEvent):
                 )
             )
             article_type = gcs_url.split("/")[-2]
-            store_article_data(file_path, article_type)
+            shop_the_look_workflow.store_article_data(file_path, article_type)
 
         print(
             f"{gcs_url} of type {e.file.mime_type} uploaded to {config.GENMEDIA_BUCKET}."
@@ -2058,305 +1908,11 @@ def on_click_upload_image(e: me.UploadEvent):
         yield
 
     if e.key == "model":
-        state.models = load_model_data()
+        state.models = shop_the_look_workflow.load_model_data()
         yield
     else:
-        load_article_data()
+        shop_the_look_workflow.load_article_data()
         yield
-
-
-def get_csv_headers(csv_reader):
-    """
-    Retrieves a list of header names from a CSV file.
-
-    Args:
-        filepath (str): The path to the CSV file.
-
-    Returns:
-        list: A list containing the header names, or an empty list if the file is empty or an error occurs.
-    """
-    try:
-
-        header = next(csv_reader)
-        return header
-    except Exception as e:
-        print(f"An error occurred: {e}")
-        return []
-
-
-db = FirebaseClient(database_id=config.GENMEDIA_FIREBASE_DB).get_client()
-
-
-def on_click_upload_models(e: me.UploadEvent):
-    """Upload image to GCS"""
-    state = me.state(PageState)
-    state.reference_model_file = e.file
-    contents = e.file.getvalue()
-    destination_blob_name = store_to_gcs(
-        "uploads", e.file.name, e.file.mime_type, contents
-    )
-
-    state.reference_model_file_gs_uri = f"gs://{destination_blob_name}"
-
-    print(
-        f"{destination_blob_name} with contents len {len(contents)} of type {e.file.mime_type} uploaded to {config.GENMEDIA_BUCKET}."
-    )
-
-    csv_file = download_from_gcs_as_string(
-        f"gs://{config.GENMEDIA_BUCKET}/uploads/{e.file.name}"
-    )
-
-    cf = [row.decode("utf-8") for row in csv_file.split(b"\n") if row]
-    cf = csv.reader(cf, delimiter=",")
-
-    required_fields = [
-        "model_group",
-        "model_id",
-        "model_name",
-        "model_description",
-        "model_view",
-        "primary_view",
-        "model_image",
-    ]
-
-    headers = get_csv_headers(cf)
-
-    for c in required_fields:  # ie. ["batch", "department"]
-        if c not in headers:
-            print(f"Missing CSV header for {c}")
-            return
-
-    current_datetime = datetime.datetime.now()
-
-    for row in cf:
-        try:
-            # TODO mapping object instead of row[]
-            doc_ref = db.collection(config.GENMEDIA_VTO_MODEL_COLLECTION_NAME).document(
-                f"{row[1]}_{row[4]}"
-            )
-            doc_ref.set(
-                {
-                    "model_group": row[0],
-                    "model_id": row[1],
-                    "model_name": row[2],
-                    "model_description": row[3],
-                    "model_view": row[4],
-                    "primary_view": row[5],
-                    "model_image": row[6],
-                    "timestamp": current_datetime,  # alt: firestore.SERVER_TIMESTAMP
-                }
-            )
-        except:
-            print(f"{row[2]} cannot be converted")
-
-
-def on_click_upload_catalog(e: me.UploadEvent):
-    """Upload image to GCS"""
-    state = me.state(PageState)
-    state.reference_catalog_file = e.file
-    contents = e.file.getvalue()
-    destination_blob_name = store_to_gcs(
-        "uploads", e.file.name, e.file.mime_type, contents
-    )
-
-    state.reference_catalog_file_gs_uri = f"gs://{destination_blob_name}"
-
-    print(
-        f"{destination_blob_name} with contents len {len(contents)} of type {e.file.mime_type} uploaded to {config.GENMEDIA_BUCKET}."
-    )
-
-    csv_file = download_from_gcs_as_string(
-        f"gs://{config.GENMEDIA_BUCKET}/uploads/{e.file.name}"
-    )
-
-    cf = [row.decode("utf-8") for row in csv_file.split(b"\n") if row]
-    cf = csv.reader(cf, delimiter=",")
-
-    required_fields = [
-        "item_id",
-        "look_id",
-        "article_type",
-        "article_color",
-        "model_group",
-        "description",
-        "image_view",
-        "try_on_order",
-    ]
-
-    headers = get_csv_headers(cf)
-
-    for c in required_fields:  # ie. ["batch", "department"]
-        if c not in headers:
-            print(f"Missing CSV header for {c}")
-            return
-
-    current_datetime = datetime.datetime.now()
-
-    for row in cf:
-        try:
-            doc_ref = db.collection(
-                config.GENMEDIA_VTO_CATALOG_COLLECTION_NAME
-            ).document(f"{row[1]}_{row[2]}")
-            doc_ref.set(
-                {
-                    "item_id": row[0],
-                    "look_id": int(row[1]),
-                    "article_type": row[2],
-                    "article_color": row[3],
-                    "model_group": row[4],
-                    "description": row[5],
-                    "image_view": row[6],
-                    "try_on_order": row[7],
-                    "timestamp": current_datetime,  # alt: firestore.SERVER_TIMESTAMP
-                }
-            )
-        except:
-            print(f"{row[2]} cannot be converted")
-
-
-def load_model_data(limit: int = 50):
-    try:
-
-        # models = []
-        # model_counter = 1
-        # folder_files = list_files_in_bucket(
-        #     f"{config.GENMEDIA_BUCKET}", prefix="uploads/models/"
-        # )
-        # for file_name in folder_files:
-        #     if file_name.split(".")[-1].lower() in ["png", "jpg", "jpeg"]:
-        #         model_counter += 1
-        #         models.append(
-        #             ModelRecord(
-        #                 model_group="4",
-        #                 model_id=f"{file_name}".split("/")[-1],
-        #                 model_name=f"{model_counter}",
-        #                 primary_view="1",
-        #                 model_image=f"{file_name}".split("/")[-1],
-        #             )
-        #         )
-        app_state = me.state(AppState)
-        media_ref = db.collection(config.GENMEDIA_VTO_MODEL_COLLECTION_NAME)
-        # .order_by(
-        #     "model_id", direction=firestore.Query.ASCENDING
-        # )
-
-        query = media_ref.where("upload_user", "in", ["everyone", app_state.user_email])
-
-        models = []
-        for doc in query.stream():
-            model_data = doc.to_dict()
-            models.append(ModelRecord(**model_data))
-
-        return models
-    except Exception as e:
-        print(f"Error fetching models: {e}")
-
-
-def load_article_data(limit: int = 50):
-    state = me.state(PageState)
-
-    # articles = []
-    # article_counter = 1
-
-    # folder_files = list_files_in_bucket(
-    #     f"{config.GENMEDIA_BUCKET}", prefix="uploads/apparel/"
-    # )
-    # for file_name in folder_files:
-    #     if file_name.split(".")[-1].lower() in ["png", "jpg", "jpeg"]:
-    #         apparel_with_subfolder = (
-    #             f"{file_name}".split("/")[-2] + "/" + f"{file_name}".split("/")[-1]
-    #         )
-    #         article_counter += 1
-    #         articles.append(
-    #             CatalogRecord(
-    #                 item_id=apparel_with_subfolder,
-    #                 look_id=0,
-    #                 article_type=f"{file_name}".split("/")[-2],
-    #                 article_color=None,
-    #                 model_group="0",
-    #                 description=None,
-    #                 ai_description=None,
-    #                 image_view="front",
-    #                 try_on_order=0,
-    #                 clothing_image_path=f"gs://{config.GENMEDIA_BUCKET}/{file_name}",
-    #             )
-    #         )
-    # state.articles = articles
-
-    app_state = me.state(AppState)
-    media_ref = db.collection(config.GENMEDIA_VTO_CATALOG_COLLECTION_NAME)
-    # .order_by(
-    #     "article_type", direction=firestore.Query.ASCENDING
-    # )
-    query = media_ref.where("upload_user", "in", ["everyone", app_state.user_email])
-
-    articles = []
-    for doc in query.stream():
-        article_data = doc.to_dict()
-        articles.append(CatalogRecord(**article_data))
-
-    articles = sorted(articles, key=lambda article: article.article_type)
-
-    state.articles = articles
-
-
-def load_look_data(limit: int = 50):
-    state = me.state(PageState)
-    # try:
-    media_ref = db.collection(config.GENMEDIA_VTO_CATALOG_COLLECTION_NAME).order_by(
-        "look_id", direction=firestore.Query.ASCENDING
-    )
-    looks = []
-    for doc in media_ref.stream():
-        # looks.append(doc.to_dict())
-        catalog_data = doc.to_dict()
-        record = CatalogRecord(**catalog_data)
-        record.clothing_image = (f"gs://{config.GENMEDIA_BUCKET}/{record.item_id}",)
-        looks.append(record)
-
-    looks.sort(key=lambda item: (item.look_id, item.try_on_order))
-    # TODO make this filter a configuration list instead of nesting way down here
-    looks = list(
-        filter(
-            lambda look: look.article_type not in ("sunglasses", "watch", "hat"),
-            looks,
-        )
-    )
-
-    return looks
-
-
-def get_selected_look():
-    state = me.state(PageState)
-    selected_look_data = list(
-        filter(lambda catalogrecord: catalogrecord.selected, state.articles)
-    )
-    # selected_look_data.sort(key=lambda item: item.try_on_order)
-    # for i in selected_look_data:
-    #     i.clothing_image = f"gs://{config.GENMEDIA_BUCKET}/{i.clothing_image}"
-
-    return selected_look_data
-
-
-# def get_selected_look(look_id):
-#     state = me.state(PageState)
-#     selected_look_data = list(
-#         filter(lambda catalogrecord: catalogrecord.look_id == look_id, state.catalog)
-#     )
-#     print(f"selected_look_data for {look_id}: {selected_look_data}")
-#     selected_look_data.sort(key=lambda item: item.try_on_order)
-
-#     return selected_look_data
-
-
-# TODO change logic to filter function
-def get_model_records(model_id):
-    state = me.state(PageState)
-    model_records = []
-    for m in state.models:
-        if m.model_id == model_id:
-            model_records.append(m)
-    return model_records
 
 
 def on_click_vto_look(e: me.ClickEvent):  # pylint: disable=unused-argument
@@ -2371,7 +1927,7 @@ def on_click_vto_look(e: me.ClickEvent):  # pylint: disable=unused-argument
     state.timing = ""  # Clear previous timing
     yield
 
-    look_articles = get_selected_look()
+    look_articles = shop_the_look_workflow.get_selected_look()
     articles_for_vto = []
 
     if e.key == "primary":
@@ -2408,7 +1964,9 @@ def on_click_vto_look(e: me.ClickEvent):  # pylint: disable=unused-argument
         articles_for_vto = look_articles
 
     print(f"articles_for_vto {articles_for_vto}")
-    images_to_process = get_model_records(state.selected_model.model_id)
+    images_to_process = shop_the_look_workflow.get_model_records(
+        state.selected_model.model_id
+    )
 
     if e.key == "alternate":
         status_prefix = "Alt View: "
@@ -2459,7 +2017,6 @@ def on_click_vto_look(e: me.ClickEvent):  # pylint: disable=unused-argument
             futures_vto = []
 
             if e.key == "primary" or e.key == "retry":
-
                 futures.append(
                     executor_catalog_enrichment.submit(
                         describe_images_and_look, look_articles
@@ -2767,293 +2324,6 @@ def on_click_vto_look(e: me.ClickEvent):  # pylint: disable=unused-argument
     vto_end_time = time.time()
     vto_duration = vto_end_time - vto_start_time
     print(vto_duration)
-
-
-# def on_click_vto_look_synchronous(e: me.ClickEvent):  # pylint: disable=unused-argument
-#     """Veo generate request handler"""
-#     vto_start_time = time.time()
-#     state = me.state(PageState)
-#     state.tryon_started = True
-#     state.is_loading = True
-#     state.show_error_dialog = False  # Reset error state before starting
-#     state.error_message = ""
-#     # state.result_image = ""  # Clear previous result
-#     state.timing = ""  # Clear previous timing
-#     yield
-
-#     look_articles = get_selected_look()
-#     articles_for_vto = []
-#     articles = []
-#     article_image_bytes_list = []
-#     print(f"look_articles is {look_articles}")
-#     for i, row in enumerate(look_articles):
-#         articles.append(row.clothing_image_path)
-
-#     # Create a list of failed items to be used later
-#     if e.key == "retry":
-#         # reverse order during retry bc try on order may have been reason for failure
-#         look_articles.reverse()
-
-#         failed_articles = list(
-#             filter(
-#                 lambda critic_record: critic_record.accurate == False,
-#                 state.final_critic.image_accuracy,
-#             )
-#         )
-
-#         failed_article_paths = []
-#         for f in failed_articles:
-#             failed_article_paths.append(f.article_image_path.split("/")[-1])
-
-#         for i, row in enumerate(look_articles):
-#             if row.item_id in failed_article_paths:
-#                 articles_for_vto.append(row)
-#     else:
-#         articles_for_vto = look_articles
-
-#     images_to_process = get_model_records(state.selected_model.model_name)
-
-#     if e.key == "alternate":
-#         status_prefix = "Alt View: "
-#     elif e.key == "retry":
-#         status_prefix = "Critic Retry: "
-#     else:
-#         status_prefix = "Primary View: "
-
-#     article_image_bytes_list_wrapper = []
-#     article_image_uri_list = []
-
-#     for r in images_to_process:
-#         # loop through images for the model (primary + alternate model)
-
-#         # TODO change from primary being int to bool
-#         if e.key == "primary" and r.primary_view != "1":
-#             continue
-#         elif e.key == "alternate" and r.primary_view == "1":
-#             continue
-#         elif e.key == "retry" and r.primary_view != "1":
-#             continue
-
-#         if e.key != "retry":
-#             state.reference_image_gcs_model = (
-#                 f"gs://{config.GENMEDIA_BUCKET}/uploads/models/{r.model_image}"
-#             )
-#         result = describe_images_and_look(look_articles)
-#         state.look_description = result.look_description
-#         for item in state.articles:
-#             for article in result.articles:
-#                 if (
-#                     item.item_id.split("/")[-1]
-#                     == article.article_image_path.split("/")[-1]
-#                 ):
-#                     item.ai_description = article.article_description
-#         yield
-
-#         for a in articles:
-#             results = download_from_gcs(a)
-#             article_image_bytes_list.append(results)
-#             # for result in results:
-#             #     article_image_bytes_list.append(result)
-
-#         for i, row in enumerate(articles_for_vto):
-#             start_time = time.time()  # Record the starting time
-
-#             # try:
-#             potential_images = []
-#             progressions = ProgressionImages()
-#             temp_progressions = []
-
-#             article_image_uri_list.append(row.clothing_image_path)
-
-#             state.current_status = f"{status_prefix}Trying on {row.article_type}..."
-#             yield
-#             # ###########THREAD 2###########
-#             # STEP 2 CALL VTO
-#             op = call_virtual_try_on(
-#                 person_image_bytes=None,
-#                 product_image_bytes=None,
-#                 person_image_uri=state.reference_image_gcs_model,
-#                 product_image_uri=row.clothing_image_path,
-#                 sample_count=int(state.vto_sample_count),
-#                 # TODO testing with these
-#                 prompt="the shirt should be tucked in",
-#                 product_description="tucked in shirt",
-#             )
-
-#             for p in op.predictions:
-#                 potential_images.append(p["gcsUri"])
-#                 temp_progressions.append(
-#                     # TODO just attach progression image instead of generating new object
-#                     ProgressionImage(
-#                         image_path=p["gcsUri"],
-#                         best_image=False,
-#                         reasoning="",
-#                     )
-#                 )
-#             # STEP 4 DOWNLOAD PROGRESSION IMAGES FROM STORAGE
-#             step_start_time = time.time()
-#             yield WorkflowStepResult(
-#                 step_name="download_candidate_images",
-#                 status="processing",
-#                 message=f"Step 1 of 7: Downloading {len(potential_images)} candidate images...",
-#                 duration_seconds=0,
-#                 data={},
-#             )
-#             reference_image_bytes_list = []
-#             with concurrent.futures.ThreadPoolExecutor() as executor:
-#                 reference_image_bytes_list = list(
-#                     executor.map(download_from_gcs, potential_images)
-#                 )
-#             step_duration = time.time() - step_start_time
-
-#             yield WorkflowStepResult(
-#                 step_name="download_candidate_images",
-#                 status="complete",
-#                 message="Reference images downloaded.",
-#                 duration_seconds=step_duration,
-#                 data={},
-#             )
-#             state.current_status = (
-#                 f"{status_prefix}Selecting best image of {row.article_type}..."
-#             )
-#             yield
-#             # STEP 5 CALL PROGRESSION CRITIC - CHOOSE BEST IMAGE
-#             step_start_time = time.time()
-#             yield WorkflowStepResult(
-#                 step_name="display_images",
-#                 status="processing",
-#                 message=f"Step 3 of 7: Display images",
-#                 duration_seconds=0,
-#                 data={},
-#             )
-#             print(article_image_bytes_list[i])
-#             best_match = select_best_image_with_description(
-#                 [article_image_bytes_list[i]],
-#                 reference_image_bytes_list,
-#                 potential_images,
-#                 f"a {row.article_type}",
-#                 f"the {row.article_type}",
-#             )
-
-#             last_best_image = None
-
-#             for p in temp_progressions:
-#                 for bm in best_match.image_accuracy:
-#                     if bm.best_image:
-#                         last_best_image = p.image_path
-#                     if bm.article_image_path == p.image_path:
-#                         p.best_image = bm.best_image
-#                         p.reasoning = bm.reasoning
-#                         p.article_image_path = bm.article_image_path
-#                         p.reasoning = bm.reasoning
-#                         p.best_image = bm.best_image
-#                         p.accurate = bm.accurate
-#                         state.result_images.append(p.image_path)
-#                         continue
-
-#             # If all images are poor, the critic refuses to select one, so select the last image to move forward with
-#             if last_best_image is None:
-#                 last_best_image = state.result_images[-1]
-#             progressions.progression_images = temp_progressions
-
-#             if e.key == "retry":
-#                 state.retry_progression_images.append(progressions)
-#             elif r.primary_view == "1":
-#                 state.progression_images.append(progressions)
-#             else:
-#                 state.alternate_progression_images.append(progressions)
-
-#             if r.primary_view == "1" and (i + 1) == len(look_articles):
-#                 state.result_image = last_best_image  # state.result_images[-1]
-#             elif i == len(look_articles):
-#                 state.alternate_images.append(p["gcsUri"])
-
-#             # TODO I believe this is just used for veo call. replace it with result_image?
-#             state.reference_image_gcs_model = last_best_image  # state.result_images[-1]
-#             yield
-
-#             # Catch specific exceptions you anticipate
-#             # except ValueError as err:
-#             #     print(f"ValueError caught: {err}")
-#             #     current_error_message = f"Input Error: {err}"
-#             # except requests.exceptions.HTTPError as err:
-#             #     print(f"HTTPError caught: {err}")
-#             #     current_error_message = f"Network/API Error: {err}"
-#             # # Catch any other unexpected exceptions
-#             # except Exception as err:
-#             #     print(f"Generic Exception caught: {type(err).__name__}: {err}")
-#             #     current_error_message = f"An unexpected error occurred: {err}"
-
-#             # finally:
-#             end_time = time.time()  # Record the ending time
-#             execution_time = end_time - start_time  # Calculate the elapsed time
-#             print(
-#                 f"Execution time: {execution_time} seconds"
-#             )  # Print the execution time
-#             state.timing = f"Generation time: {round(execution_time)} seconds"
-#             state.current_status = ""
-#             yield
-
-#     print("Cut! That's a wrap!")
-
-#     if e.key != "retry":
-#         if state.generate_alternate_views and e.key == "primary":
-#             new_event = e
-#             new_event.key = "alternate"
-#             yield from on_click_vto_look(new_event)
-#         elif state.generate_video:
-#             video_result = on_click_veo(e)
-#             yield from on_click_veo(e)
-
-#     if e.key == "primary" or e.key == "retry":
-#         step_start_time = time.time()
-#         yield WorkflowStepResult(
-#             step_name="final_critic",
-#             status="processing",
-#             message=f"Step 9000 of 7: final critic",
-#             duration_seconds=0,
-#             data={},
-#         )
-
-#         with concurrent.futures.ThreadPoolExecutor() as executor:
-#             final_image_bytes_list = list(
-#                 executor.map(download_from_gcs, [state.result_image])
-#             )
-#             state.current_status = "Critic evaluation in progress..."
-#             yield
-
-#             final_critic = final_image_critic(
-#                 # article_image_bytes_list_wrapper,
-#                 article_image_bytes_list,
-#                 article_image_uri_list,
-#                 final_image_bytes_list,
-#             )
-#             state.final_critic = final_critic
-#             state.final_accuracy = final_critic.accurate
-#             state.current_status = ""
-#             yield
-#             # minus-circle alert-decagram
-#             if e.key == "primary" and not state.final_critic.accurate:
-#                 # retry logic
-#                 new_event = e
-#                 new_event.key = "retry"
-#                 yield from on_click_vto_look_synchronous(new_event)
-
-#         step_duration = time.time() - step_start_time
-#         yield WorkflowStepResult(
-#             step_name="final_critic",
-#             status="complete",
-#             message="Article description generated",
-#             duration_seconds=step_duration,
-#             data={},
-#         )
-
-#     state.is_loading = False
-#     yield
-
-#     vto_end_time = time.time()
-#     vto_duration = vto_end_time - vto_start_time
-#     print(vto_duration)
 
 
 def workflows_content_retail_look():
